+++ apps/observations/forms.py
@@ -4,62 +4,6 @@ from django.utils.translation import gettext_lazy as _
 from base.forms import FilterForm
 
 
-class ObservationFilter(FilterForm):
-    form_id = "observation"
-    form_title = _("Options")
-
-    approved = forms.ChoiceField(
-        label=_("Status"),
-        choices=(
-            ("", _("All")),
-            ("True", _("Accepted")),
-            ("False", _("Rejected")),
-        ),
-        required=False,
-        widget=forms.Select(attrs={"class": "form-control"}),
-    )
-
-    audio = forms.ChoiceField(
-        label=_("With audio"),
-        choices=(
-            ("", _("All")),
-            ("True", _("Yes")),
-            ("False", _("No")),
-        ),
-        required=False,
-        widget=forms.Select(attrs={"class": "form-control"}),
-    )
-
-    photo = forms.ChoiceField(
-        label=_("With photos"),
-        choices=(
-            ("", _("All")),
-            ("True", _("Yes")),
-            ("False", _("No")),
-        ),
-        required=False,
-        widget=forms.Select(attrs={"class": "form-control"}),
-    )
-
-    video = forms.ChoiceField(
-        label=_("With video"),
-        choices=(
-            ("", _("All")),
-            ("True", _("Yes")),
-            ("False", _("No")),
-        ),
-        required=False,
-        widget=forms.Select(attrs={"class": "form-control"}),
-    )
-
-    filters = {
-        "approved": "approved",
-        "audio": "audio",
-        "photo": "photo",
-        "video": "video",
-    }
-
-
 class ObservationOrder(FilterForm):
     form_id = "observation-order"
     form_title = _("Order By")